Consistency with Windows
On POSIX systems, the rename system call will silently replace the destination file if the user has sufficient permissions. The same is not true on Windows.

os.replace and os.rename are the same function on POSIX systems, but on Windows os.replace will call MoveFileExW with the MOVEFILE_REPLACE_EXISTING flag set to give the same effect as on POSIX systems.