Remove common prefixes from filenames (#45)
I removed common prefixes from filenames. This allows to verify
filenames like "1.2.86.1.dcm", ..., "1.2.86.12.dcm".

I also removed `vdcms = [d.rsplit('.', 1)[0] for d in filenames]` since
this is redundant in combination with the line after it: `vdcms =
[int(''.join(c for c in str(fn) if c.isdigit())) for fn in filenames]`

Finally, `True` is returned if filenames could not be verified for a
particular series: only one slice or no number in the filename. This
makes the default of checking for numbers more user friendly.